changing roaringbitvec interface to size_t

+++ libcocadastrproc/src/roaring.h
@@ -22,6 +22,7 @@
 #ifndef ROARING_H
 #define ROARING_H
 
+#include <stddef.h>
 #include <stdlib.h>
 #include <stdio.h>
 
@@ -41,30 +42,29 @@ roaringbitvec *roaringbitvec_new_from_bitarr(byte_t *b, size_t n);
 
 void roaringbitvec_free(roaringbitvec *self);
 
-uint32_t roaringbitvec_card(roaringbitvec *self);
+size_t roaringbitvec_card(roaringbitvec *self);
 
-uint32_t roaringbitvec_count(roaringbitvec *self, bool bit);
+size_t roaringbitvec_count(roaringbitvec *self, bool bit);
 
 size_t roaringbitvec_memsize(roaringbitvec *self);
 
 void roaringbitvec_fit(roaringbitvec *self);
 
-void roaringbitvec_set(roaringbitvec *self, uint32_t pos, bool val);
+void roaringbitvec_set(roaringbitvec *self, size_t pos, bool val);
 
-bool roaringbitvec_get(roaringbitvec *self, uint32_t pos);
+bool roaringbitvec_get(roaringbitvec *self, size_t pos);
 
-uint32_t roaringbitvec_rank0(roaringbitvec *self, uint32_t pos);
+size_t roaringbitvec_rank0(roaringbitvec *self, size_t pos);
 
-uint32_t roaringbitvec_rank1(roaringbitvec *self, uint32_t pos);
+size_t roaringbitvec_rank1(roaringbitvec *self, size_t pos);
 
-uint32_t roaringbitvec_rank(roaringbitvec *self, bool bit, uint32_t pos);
+size_t roaringbitvec_rank(roaringbitvec *self, bool bit, size_t pos);
 
-uint32_t roaringbitvec_select0(roaringbitvec *self, uint32_t rank);
+size_t roaringbitvec_select0(roaringbitvec *self, size_t rank);
 
-uint32_t roaringbitvec_select1(roaringbitvec *self, uint32_t rank);
+size_t roaringbitvec_select1(roaringbitvec *self, size_t rank);
 
-uint32_t roaringbitvec_select(roaringbitvec *self, bool bit, uint32_t rank);
-
+size_t roaringbitvec_select(roaringbitvec *self, bool bit, size_t rank);
 
 void roaringbitvec_fprint(FILE *stream, roaringbitvec *self);
